requirements.txt
flask
sqlalchemy
flask-sqlalchemy
pymysql
python-dotenv
jinja2
openpyxl
cryptography
email-validator
